Prevented `size(self)` from downloading data
For `WikipediaDataset` and `GutenbergDataset` the code silently downloads the files in question and then directly measures their sizes. These have been replaced by placeholder values of 1 to prevent this behavior. It may be desirable to define `size(self)` and `num_docs(self)` in the abstract class and have them check to see if there exists a dataset of the right name, and empirically measures it if it is there. Otherwise, raise an error.